requirements.txt
transformers 
python-dotenv 
langchain 
langchain-huggingface
PyPDF2
python-docx